Refactoring: searchInput
